+++ src/test/java/com/solace/spring/cloud/stream/binder/inbound/queue/FlowXMLMessageListenerTest.java
@@ -1,15 +1,14 @@
 package com.solace.spring.cloud.stream.binder.inbound.queue;
 
+import com.solace.spring.cloud.stream.binder.util.WatchdogLogger;
 import com.solacesystems.jcsmp.BytesXMLMessage;
 import com.solacesystems.jcsmp.JCSMPFactory;
 import com.solacesystems.jcsmp.Topic;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
-import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
@@ -22,15 +21,14 @@ class FlowXMLMessageListenerTest {
 
     @Test
     void testStartReceiverThreads_StartsSpecifiedNumberOfThreads() throws InterruptedException {
-        FlowXMLMessageListener listener = new FlowXMLMessageListener();
+        FlowXMLMessageListener listener = new FlowXMLMessageListener(new WatchdogLogger());
         Consumer<BytesXMLMessage> messageConsumer = Mockito.mock(Consumer.class);
 
         int threadCount = 3;
         String threadNamePrefix = "testStartReceiverThreads_StartsSpecifiedNumberOfThreads";
-        long processingTime = 1000;
 
         // Start the receiver threads
-        listener.startReceiverThreads(threadCount, threadNamePrefix, messageConsumer, processingTime);
+        listener.startReceiverThreads(threadCount, threadNamePrefix, messageConsumer, 3, 300, 2000);
 
         // Wait briefly to let threads initialize
         Thread.sleep(1000);
@@ -48,15 +46,14 @@ class FlowXMLMessageListenerTest {
 
     @Test
     void testStartReceiverThreads_CallsMessageConsumerWhenMessageIsPolled() throws InterruptedException {
-        FlowXMLMessageListener listener = new FlowXMLMessageListener();
+        FlowXMLMessageListener listener = new FlowXMLMessageListener(new WatchdogLogger());
         Consumer<BytesXMLMessage> messageConsumer = Mockito.mock(Consumer.class);
 
         int threadCount = 1;
         String threadNamePrefix = "TestThread";
-        long processingTime = 1000;
 
         // Start the receiver threads
-        listener.startReceiverThreads(threadCount, threadNamePrefix, messageConsumer, processingTime);
+        listener.startReceiverThreads(threadCount, threadNamePrefix, messageConsumer, 3, 300, 2000);
 
         // Simulate a message being received
         BytesXMLMessage mockMessage = mock(BytesXMLMessage.class);
@@ -74,7 +71,7 @@ class FlowXMLMessageListenerTest {
 
     @Test
     void testStartReceiverThreads_useAllThreads() throws InterruptedException {
-        FlowXMLMessageListener listener = new FlowXMLMessageListener();
+        FlowXMLMessageListener listener = new FlowXMLMessageListener(new WatchdogLogger());
         Consumer<BytesXMLMessage> messageConsumer = Mockito.mock(Consumer.class);
         List<BytesXMLMessage> results = new ArrayList<>();
         doAnswer(invocation -> {
@@ -89,10 +86,9 @@ class FlowXMLMessageListenerTest {
 
         int threadCount = 10;
         String threadNamePrefix = "testStartReceiverThreads_useAllThreads";
-        long processingTime = 1000;
 
         // Start the receiver threads
-        listener.startReceiverThreads(threadCount, threadNamePrefix, messageConsumer, processingTime);
+        listener.startReceiverThreads(threadCount, threadNamePrefix, messageConsumer, 3, 300, 2000);
 
         // Wait briefly to let threads initialize
         Thread.sleep(1000);
@@ -118,38 +114,50 @@ class FlowXMLMessageListenerTest {
     }
 
     @Test
-    void testStartReceiverThreads_WatchdogLogsWarningForLongProcessing() throws NoSuchFieldException, IllegalAccessException {
-        FlowXMLMessageListener listener = new FlowXMLMessageListener();
-        // Use reflection to access the private 'activeMessages' field
-        Field activeMessagesField = FlowXMLMessageListener.class.getDeclaredField("activeMessages");
-        activeMessagesField.setAccessible(true);
-        Set<FlowXMLMessageListener.MessageInProgress> activeMessages = (Set<FlowXMLMessageListener.MessageInProgress>) activeMessagesField.get(listener);
+    void testStartReceiverThreads_WatchdogLogsWarningForCongestedQueue() {
+        WatchdogLogger mockWatchdogLogger = Mockito.mock(WatchdogLogger.class);
+        doCallRealMethod().when(mockWatchdogLogger).warnIfNecessary(any(), any(), any(), any());
+        doCallRealMethod().when(mockWatchdogLogger).setLatestWarning(any());
+
+        FlowXMLMessageListener listener = new FlowXMLMessageListener(mockWatchdogLogger);
+
         Consumer<BytesXMLMessage> messageConsumer = message -> {
             try {
                 // Simulate a long message processing time
-                Thread.sleep(6000);
+                Thread.sleep(8000);
             } catch (InterruptedException ignored) {
             }
         };
 
-        int threadCount = 1;
+        int threadCount = 2;
         String threadNamePrefix = "WatchdogTestThread";
-        long processingTime = 500;
 
         // Start the receiver threads
-        listener.startReceiverThreads(threadCount, threadNamePrefix, messageConsumer, processingTime);
+        listener.startReceiverThreads(threadCount, threadNamePrefix, messageConsumer, 3, 300, 2000);
 
-        // Simulate a message being received
-        BytesXMLMessage mockMessage = mock(BytesXMLMessage.class);
-        Mockito.when(mockMessage.getMessageId()).thenReturn("TestMessageId");
-        Mockito.when(mockMessage.getDestination()).thenReturn(JCSMPFactory.onlyInstance().createTopic("test/topic"));
-        listener.onReceive(mockMessage);
+        // Simulate messages being received
+        Topic topic = JCSMPFactory.onlyInstance().createTopic("test/topic");
+        for (int i = 0; i < 6; i++) {
+            BytesXMLMessage mockMessage = mock(BytesXMLMessage.class);
+            Mockito.when(mockMessage.getMessageId()).thenReturn("TestMessageId");
+            Mockito.when(mockMessage.getDestination()).thenReturn(topic);
+            listener.onReceive(mockMessage);
+        }
 
+        mockWatchdogLogger.setLatestWarning(System.currentTimeMillis() - (6 * 60 * 1000));
 
-        // Wait for the message to be marked as warned in the activeMessages map
-        await().atMost(700, TimeUnit.MILLISECONDS)
-                .until(() -> activeMessages.iterator().next().isWarned());
-        await().atMost(5500, TimeUnit.MILLISECONDS)
-                .until(() -> activeMessages.iterator().next().isErrored());
+        // Wait for the warning to be logged
+        verify(mockWatchdogLogger, timeout(8000).times(1)).logRelaxed(any(), any());
+
+        for (int i = 0; i < 10; i++) {
+            BytesXMLMessage mockMessage = mock(BytesXMLMessage.class);
+            Mockito.when(mockMessage.getMessageId()).thenReturn("TestMessageId");
+            Mockito.when(mockMessage.getDestination()).thenReturn(topic);
+            listener.onReceive(mockMessage);
+        }
+
+        mockWatchdogLogger.setLatestWarning(System.currentTimeMillis() - (6 * 60 * 1000));
+
+        verify(mockWatchdogLogger, timeout(8000).times(1)).logUrgent(any(), any(), any());
     }
 }